Added ability to change feed and folder names

+++ client/src/components/EditModal.js
@@ -5,16 +5,47 @@ function EditModal({ name, id, onClose}) {
     const [nameInput, setNameInput] = useState(name);
     const { updateSubscriptions } = useSubscriptions();
     
-    function handleSubmit() {
-        
-        return null;
+    const isFeed = (id !== undefined);
+    
+    async function handleSubmit(e) {
+        e.preventDefault();
+        if (isFeed) {
+            await fetch('/subscriptions/rename', {
+                method: 'PATCH',
+                headers: {
+                    'Content-Type': 'application/json'
+                },
+                body: JSON.stringify({
+                    subscriptionid: id,
+                    newName: nameInput
+                })
+            });
+        } else {
+            await fetch('/subscriptions/rename-folder', {
+                method: 'PATCH',
+                headers: {
+                    'Content-Type': 'application/json'
+                },
+                body: JSON.stringify({
+                    oldName: name,
+                    newName: nameInput
+                })           
+            });
+        } 
+        updateSubscriptions();
+        onClose();
     }
     
     async function handleDelete(e) {
         e.preventDefault();
-        await fetch(`/subscriptions?subscriptionid=${id}`, {
-            method: 'DELETE'
-        });
+        if (isFeed) {
+            await fetch(`/subscriptions?subscriptionid=${id}`, {
+                method: 'DELETE'
+            });
+        } else {
+            //todo
+            //await fetch (`/subscriptions`)
+        }
         updateSubscriptions();
         onClose();
     }
